move /users/me/ to the users-service to include auth headers
